requirements.txt
flask
python-dotenv
azure-cosmosdb-table